Split up types into separate files.
