+++ src/System.Text.JsonLab/System/Text/Json/JsonWriterException.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Text.JsonLab
+{
+    public sealed class JsonWriterException : Exception
+    {
+        public JsonWriterException(string message) : base(message)
+        {
+        }
+    }
+}
